+++ src/umleditor/mvc_controller/controller.py
@@ -1,18 +1,12 @@
 from .controller_input import read_file, read_line
 import umleditor.mvc_controller.controller_output as controller_output
-from umleditor.mvc_controller.uml_parser import Parser
+from umleditor.mvc_controller.uml_parser import parse
 from .serializer import CustomJSONEncoder, serialize, deserialize
 from umleditor.mvc_model import CustomExceptions as CE
 from umleditor.mvc_model.diagram import Diagram
 from umleditor.mvc_model import help_menu
 from umleditor.mvc_controller.uml_parser import check_args
 import os
-
-#Parser Includes. These will be moved out when the parser is moved.
-from umleditor.mvc_model.entity import Entity
-from umleditor.mvc_model.relation import Relation
-
-
 
 
 class Controller:
@@ -22,12 +16,11 @@ class Controller:
 
     
     def run(self) -> None:
-        p = Parser(self._diagram)
         while not self._should_quit:
             s = read_line()
             try:
                 #parse the command
-                input = p.parse(s)
+                input = parse(self, s)
 
                 #return from input is [function object, arg1,...,argn]
                 command = input[0]
